+++ pdtable/io/_excel_openpyxl.py
@@ -1,6 +1,6 @@
 """Machinery to read/write Tables in an Excel workbook using openpyxl as engine."""
-from typing import Union, Iterable, Sequence, Any
 from os import PathLike
+from typing import Union, Iterable, Sequence, Any, Dict
 
 import openpyxl
 
@@ -15,6 +15,9 @@ from pdtable import Table
 from pdtable.io._represent import _represent_row_elements
 
 
+DEFAULT_SHEET_NAME = "Sheet1"
+
+
 def read_cell_rows_openpyxl(path: Union[str, PathLike]) -> Iterable[Sequence[Any]]:
     """Reads from an Excel workbook, yielding one row of cells at a time."""
     import openpyxl
@@ -25,14 +28,26 @@ def read_cell_rows_openpyxl(path: Union[str, PathLike]) -> Iterable[Sequence[Any
 
 
 def write_excel_openpyxl(tables, path, na_rep):
-    """Writes tables to an Excel workbook at the specified path."""
-    if isinstance(tables, Table):
-        # For convenience, pack single table in an iterable
-        tables = [tables]
+    """Write tables to an Excel workbook at the specified path."""
+
+    if not isinstance(tables, Dict):
+        # For convenience, pack it in a dict
+        tables = {DEFAULT_SHEET_NAME: tables}
+
     wb = openpyxl.Workbook()
-    ws = wb.active
-    for t in tables:
-        _append_table_to_openpyxl_worksheet(t, ws, na_rep)
+    wb.remove(wb.active)  # Remove the one sheet that openpyxl creates by default
+
+    for sheet_name in tables:
+
+        tabs = tables[sheet_name]
+        if isinstance(tabs, Table):
+            # For convenience, pack single table in an iterable
+            tabs = [tabs]
+
+        ws = wb.create_sheet(title=sheet_name)
+        for t in tabs:
+            _append_table_to_openpyxl_worksheet(t, ws, na_rep)
+
     wb.save(path)